requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
anyio==4.6.2.post1
APScheduler==3.11.0
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
paramiko==3.5.0
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.10
pycparser==2.22
PyNaCl==1.5.0
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
redis==5.2.0
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
sshtunnel==0.4.0
tzlocal==5.2
urllib3==2.2.3
yarl==1.16.0
